Type constants as uint64 when possible

+++ src/common/CanonicalMachine.sol
@@ -15,27 +15,27 @@ library CanonicalMachine {
     uint256 constant INPUT_MAX_SIZE = 1 << EmulatorConstants.PMA_CMIO_TX_BUFFER_LOG2_SIZE;
 
     /// @notice Log2 of maximum number of outputs.
-    uint256 constant LOG2_MAX_OUTPUTS = 63;
+    uint64 constant LOG2_MAX_OUTPUTS = 63;
 
     /// @notice Log2 of memory size.
-    uint256 constant LOG2_MEMORY_SIZE = Memory.LOG2_MAX_SIZE;
+    uint64 constant LOG2_MEMORY_SIZE = Memory.LOG2_MAX_SIZE;
 
     /// @notice Log2 of Merkle tree data block size.
     /// @dev Used when computing the the machine memory Merkle root.
-    uint256 constant LOG2_DATA_BLOCK_SIZE = EmulatorConstants.TREE_LOG2_WORD_SIZE;
+    uint64 constant LOG2_DATA_BLOCK_SIZE = EmulatorConstants.TREE_LOG2_WORD_SIZE;
 
     /// @notice Merkle tree height.
     /// @dev height
     ///      = log_2 #leaves
     ///      = log_2 (memorySize / dataBlockSize)
     ///      = log_2 memorySize - log_2 dataBlockSize
-    uint256 constant TREE_HEIGHT = LOG2_MEMORY_SIZE - LOG2_DATA_BLOCK_SIZE;
+    uint64 constant TREE_HEIGHT = LOG2_MEMORY_SIZE - LOG2_DATA_BLOCK_SIZE;
 
     /// @notice The start address of the tx buffer.
     /// @dev Used to write outputs and the outputs Merkle root.
-    uint256 constant TX_BUFFER_START = EmulatorConstants.PMA_CMIO_TX_BUFFER_START;
+    uint64 constant TX_BUFFER_START = EmulatorConstants.PMA_CMIO_TX_BUFFER_START;
 
     /// @notice The leaf index of the outputs Merkle root (tx buffer start).
     /// @dev Useful for validating the outputs Merkle root from the machine root.
-    uint256 constant OUTPUTS_ROOT_LEAF_INDEX = TX_BUFFER_START >> LOG2_DATA_BLOCK_SIZE;
+    uint64 constant OUTPUTS_ROOT_LEAF_INDEX = TX_BUFFER_START >> LOG2_DATA_BLOCK_SIZE;
 }
